+++ backend/src/model/types/common.rs
@@ -64,7 +64,7 @@ pub fn contains_provider_type(
     providers.get(&provider.provider_type()).is_some()
 }
 
-pub fn calculate_fees(fiat_amount: u64, crypto_amount: u64) -> (u64, u64) {
+pub fn calculate_fees(fiat_amount: u64, crypto_amount: u128) -> (u64, u128) {
     // Static strategy: 2% fee for the offramper, 0.5% for the admin
     let offramper_fee = fiat_amount / 50; // 2%
     let crypto_fee = crypto_amount / 200; // 0.5%
